this does not work

+++ user_management/src/user_management/utils.py
@@ -4,9 +4,9 @@ import string
 from io import BytesIO
 
 import requests
+from PIL import Image
 from django.core.files import File
 from django.core.files.base import ContentFile
-from PIL import Image
 
 from user.models import User
 from user_management import settings
@@ -37,13 +37,20 @@ def download_image_from_url(url, model_instance):
 
 
 def save_image_from_base64(base64_string, model_instance):
-    img = Image.open(BytesIO(base64.b64decode(base64_string)))
-    img_io = BytesIO()
-    img.save(img_io, format='PNG')
-    img_file = ContentFile(img_io.getvalue())
-    random_suffixes = generate_random_string(10)
-    model_instance.avatar.delete()
-    model_instance.avatar.save(f'{model_instance.id}_{random_suffixes}.png', File(img_file), save=True)
+    if not base64_string:
+        return False, 'Image not found'
+    if not base64_string.startswith('data:image/png;base64,'):
+        return False, 'Invalid image format'
+    try :
+        img = Image.open(BytesIO(base64.b64decode(base64_string)))
+        img_io = BytesIO()
+        img.save(img_io, format='PNG')
+        img_file = ContentFile(img_io.getvalue())
+        random_suffixes = generate_random_string(10)
+        model_instance.avatar.delete()
+        model_instance.avatar.save(f'{model_instance.id}_{random_suffixes}.png', File(img_file), save=True)
+    except Exception as e:
+        return False, str(e)
     return True
